Add files via upload
Here is a working version of MAC valve control and using some IO for the Spartan 2 OEM Wideband controller. Currently trying to add in the ability to only turn on the MAC valve to the selected setting when it detects positive pressure on the MPX4250 MAP sensor. That way it stays off during idle and normal driving to not create a vacuum leak.